Connect/disconnect as Skype becomes available/unavailable
We can easily track the presence of Skype's service on D-Bus, but
unfortunately, when no user is signed in, we can't connect to
Skype yet.  And there's no way of knowing when a user signs in or
signs out, which is quite an annoying limitation of Skype's API.
We work around this problem by polling.  When Skype gets visible
on D-Bus, We try to connect every 5 seconds until it we get a
successful connection.  After that, we PING it every 5 seconds
until it doesn't reply to it anymore.