Fix bug with calls on hold
Previously, when explicitely stopping to record a call on hold, as
soon as the call resumes it would be considered to be recorded
again.  If it's set to auto-record, it would then silently start
recording again.

Now, CallHandler does not decide anymore when to start recording,
it now happens in Call::setStatus().

Also, this now prevents the status to be re-set on every DURATION
line.