Introduce DISABLE_COPY_AND_ASSIGNMENT macro
This is much shorter than writing it every time.  Also, some
classes were missing this so far.

It's arguable whether this makes any sense at all, but it's not
going to hurt either.  The classes that use this are never meant
to be copied or assigned to and in some cases only one instance is
supposed to exist anyway.