utils.cpp: Set FD_CLOEXEC on lock file
Also see commit ce6f838b4587528630784c1ea1ad272b64e78544.  There
was one remaining issue when Skype Call Recorder would start, get
the lock file and then fork off dbus-daemon which would inherit
the lock.  In that situation, if Skype Call Recorder is properly
terminated, the lock is released and deleted and dbus-daemon would
hold a harmles open fd to the now deleted lock file.  But when it
crashes instead or gets interrupted, it will not release the lock,
causing the problem described in that other commit.