Fix bug about preferences messing up
The reference semantics of QList<Preference> (implemented as an
array of pointers) during qSort() aren't what I expected.
Existing references will continue to point at the element at the
same index, instead of the same element at its new position.  This
is because qSort() swaps elements with temporaries and
assignments, instead of shuffling the internal pointer array.

Using QLinkedList has the best reference semantics, but is
unpractical for sorting.  So now we only store the pointers in a
QList and manage them manually.

This caused a strange bug (originally observed by Jan Martin),
where the references held by the smart widgets suddenly pointed to
other preferences after sorting, which caused them to store values
in the wrong preferences.