New debug output for studying stream synchronization
The new option debug.writesyncfile (not available in the
preferences dialog) will write out a sync file, which can be
viewed with utils/syncviewer.

The reason for this is that someone has reported a stream being
more than 5 seconds out of sync after 3 hours of recording.  It's
unclear how such delays can happen.