Delay error message when no system tray is present
When configuring a desktop environment to start Skype Call
Recording automatically on start-up, it could happen the it was
started before the environment provided a system tray, causing it
to fail with an error message.  QSystemTrayIcon explicitely allows
to create the icon before the tray becomes available as it will
then take care of adding it as soon as possible.  The error
message now appears only if the system tray is still not available
after 10 seconds.

Note that this code was buggy anyway: If no system tray was
detected, the signal requestQuitNoConfirmation() was emitted,
which did nothing because it wasn't connected to anything yet (no
way it could, since TrayIcon was just only constructed).