Add dbus-x11 as dependency and improve DBus error handling
The package dbus-x11 provides dbus-launch, which is indirectly
required to launch a DBus if none is running already.  If it
wasn't installed, Skype Call Recorder was almost silently failing
to launch DBus and to connect.  If there's now an error with
either connecting to DBus or registrating our object, an error
message is displayed.