Make mixing to mono a real mix, instead of just adding channels
One could argue whether this is a better thing to do, because
usually in a normal call, only one person is talking at a time, so
clipping is unlikely to happen.  But averaging channels is the
normal way to mix to mono.

Also, don't use a separate buffer anymore, but mix directly into
bufferLocal.