New stereo mixing
This new code replaces the old 'mono', 'stereo' and 'oerets'
modes.  It's now a switch to enable stereo and a mix preference
that allows to pan the local/remote channels arbitrarily between
the left/right channels.

A follow-up commit will add code to automatically convert old
preferences to the new format.