Add support for the Asus Eee PC
Running 'utils/makedeb -e' now builds a .deb package that is
suitable for the Asus Eee PC.  It has libmp3lame and libid3
statically linked, because the Xandros linux on the Eee doesn't
provide those.

In order to compile statically, the directory 'static/' must be
prepared to contain the libraries to build against, see the README
file.

Thanks a lot to mike <n5pwp@comcast.net>, who helped me making
this work.