Fix dangling pointer in QCoreApplication
The constructors of Q[Core]Application takes its 'argc' argument
as an 'int &' and will keep a pointer to it for as long as it
exists.  But the class Recorder, which inherits QApplication,
would pass it a reference that goes out of scope after it's been
constructed.

This surprisingly seemed to not cause problems on x86, but somehow
always caused a crash on amd64.

Thanks to Arthur Richards that provided me the info that helped me
find this bug.